client: Improve handling of server/scribe URLs.
By default, use paths on the same host (and the same protocol and
port) from which this script is loaded.  If the script is loaded from
file://, use https://physionet.org/ as the host.

The server and scribe URLs can be overridden by 'server' and 'scribe'
query parameters in the page URL.