client: use friendlier options for JSON requests.
Set 'crossDomain' to true in order to use JSONP (regardless of whether
the origin of the script matches the server URL) and avoid setting the
'X-Requested-With' header.

Set 'cache' to true in order to stop jquery from adding useless
garbage to the URL designed to prevent caching.  (It's unclear whether
this is sufficient to actually make caching *work*, but at least we
don't need to add extra roadblocks.)