server: implement a sandbox using chroot and seccomp.
This requires linux 3.5 or later (at least) to work, and libseccomp.
The server will need to be installed setuid.  It will require careful
setup of the chroot directory and WFDB path to ensure that the desired
files are visible.  The sandbox does not allow any network access.

The non-sandboxed server will now refuse to run as root, in case
somebody installs the wrong thing accidentally.