sandbox: honor the LIGHTWAVE_WFDBCAL environment variable.
If set, this must be the path (external to LIGHTWAVE_ROOT) to the
wfdbcal file.  This file will be opened before performing the chroot,
but will not be read or parsed until after chrooting and dropping
privileges.  (This is done in a somewhat kludgy way, by opening the
file as standard input and then setting the WFDBCAL environment
variable to "-"; this only works because lightwave doesn't use its
standard input for anything else.)

Like LIGHTWAVE_ROOT, the LIGHTWAVE_WFDBCAL file must be readable by
the real uid/gid.

Note that LIGHTWAVE_WFDBCAL has no effect on non-sandboxed server
configurations; use WFDBCAL instead.