client: Make JSONP optional, and use JSON by default.
The LightWAVE server can operate in two modes: JSON and JSONP.

JSON mode works only if the server is located on the same host (same
origin) as the main document, or if the server explicitly permits
cross-origin access by adding an 'Access-Control-Allow-Origin' header.
(The PhysioNet LightWAVE server does this for public databases.)

JSONP mode works in cross-domain contexts, which can be helpful, but
is also a security risk if the server is untrusted (since the response
is executed as JavaScript) or if the database is private (since the
browser permits any webpage to read the data.)

To explicitly request JSONP, add '?callback=?' when you specify the
server URL; e.g. 'https://physionet.org/cgi-bin/lightwave?callback=?'.