setrepos: let environment override LW_WFDB.
In general, the WFDB variable won't be set in the web server's
environment.  By allowing the environment variable to override
LW_WFDB, we can have a single (sandboxed-)lightwave binary whose
default path is different from the WFDB library's default path, but
the path can also be overridden if necessary.