sandbox: permit openat only with AT_FDCWD.
As noted in commit 01d123b69095fd60c65772df62725a1eb068ee5c, newer
versions of glibc implement the open() and fopen() functions using the
syscall openat(AT_FDCWD, ...).  This is equivalent to the traditional
open() syscall and is safe.

However, if the program is allowed to use an arbitrary file descriptor
as the dirfd argument to openat, and if the sandboxed-lightwave
executable is accessible in an environment that is already constrained
by chroot, and a malicious local user is able to exploit a bug in
libwfdb for code execution, this would allow them to read files that
are located outside the outer chroot.

Somewhat surprisingly, on 64-bit systems, the first argument to openat
(at least as far as seccomp is concerned) is (uint32_t)AT_FDCWD
(0xffffff9c), rather than (uint64_t)AT_FDCWD (0xffffffffffffff9c).