client: Allow traces that don't align exactly with window.
When we request data from the server, we specify the start (t0) and
duration (dt) in seconds, and the server converts these to WFDB frame
numbers; this may not correspond exactly to the start/end of the
window, especially when the record has multiple frequencies and/or has
a non-integer frame frequency.

Therefore, when drawing the plot, look for any trace that overlaps
with the window, not only traces that overlap with the exact window
starting time (t0_ticks).  Furthermore, if we don't have traces for a
particular signal that cover the entire window, don't give up
completely; just draw the traces we do have and keep going.