client: Fix plotting unaligned traces.
In the show_plot function:

- tnext is the tick number of the left edge of the area to be drawn,
  which is initally the left edge of the window (t0_ticks), and
  increases after each trace is drawn.

- tf is the tick number of the right edge of the area to be drawn,
  which is either the right edge of the window (tf_ticks), or the end
  of the record (rdt_ticks).

- t is a relative tick number (offset from t0_ticks).

- x is a relative X coordinate, which should equal t*xstep.

- imin and imax are the indices into the sample array corresponding to
  the data we want to draw.

It can be seen that the previous code for calculating imin in the
first iteration would do the same as in subsequent iterations, since
tnext is initially equal to t0_ticks.

However, the calculation of imin didn't account for the fact that the
selected trace might begin *after* tnext (which can happen if the
server rounds the starting timestamp up due to multi-frequency) and
didn't account for the fact that (tnext - trace.t0) might not be a
multiple of tps (i.e., the window might not start on a frame
boundary.)

Likewise, it can be seen that the previous code for calculating imax
in the first iteration would do the same as in subsequent iterations,
except that the first iteration wouldn't clamp to the record length
while later iterations would.  This shouldn't matter normally, but
might give odd results for incorrectly-formatted records, and it's
better to be consistent.

The calculation of t and x can be simplified, and account for windows
not starting on a frame boundary, by calculating these in terms of imin
and imax.