Add a simple server for running LightWAVE locally.
lw-server is a simple script that provides the web-server component of
LightWAVE, without any special setup or privileges necessary.  Run
this script with a local database directory as the argument (i.e., the
path to a directory containing a "RECORDS" file); this will run an
HTTP server listening on the loopback interface, which you can then
access with a web browser.

This uses the standard http.server module of Python 3, which is not
designed for security (or performance) and shouldn't be used in a
production system, but is simple, widely available, and adequate for
running LightWAVE locally on trusted data.

(For viewing untrusted data, using sandboxed-lightwave is recommended;
lw-server does not currently support sandboxed-lightwave.)