sandbox: allow running unprivileged using a user namespace.
Instead of installing sandboxed-lightwave as a setuid executable, we
can use a user namespace as a way to gain privileges to perform
chroot.  (This trades off one theoretical attack surface for another.)

Note that user namespaces require Linux 3.8 or later, and historically
many distributions have disabled them or restricted their use.
Unprivileged mode also won't work if running in an environment that is
already restricted by chroot (which is not the same thing as running
in a full container - I think unprivileged mode should work inside a
container, provided that nested namespaces are allowed.)