sandbox: fix display of values starting with "0x1".
The sandbox provides a handler for SIGSYS so that if a system call is
forbidden by the seccomp filter, a useful error message is printed to
standard error.  The nature of trying to debug seccomp means this
handler needs to use the bare minimum of system calls, so instead of
using library functions like sprintf or dprintf, we output numbers "by
hand", one digit at a time.  This logic was incorrect if the first
digit was a "1".